feat: add BaseEntity, BaseTimeEntity
