--- src/main/java/WOWMATE/Server/ServerApplication.java
@@ -1,13 +1,0 @@
-package WOWMATE.Server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ServerApplication.class, args);
-	}
-
-}
