--- src/main/java/WOWMATE/Server/test.java
@@ -1,6 +1,0 @@
-package WOWMATE.Server;
-
-public class test {
-
-
-}
