--- src/main/java/chatroom/domain/Message.java
@@ -1,25 +1,0 @@
-package chatroom.domain;
-
-import chatroom.domain.Chatroom;
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Setter
-public class Message {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "message_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chatroom_id")
-    private Chatroom chatRoom;
-
-    private Long sendUserId;
-    private String content;
-}
