--- src/main/java/chatroom/repository/ChatroomRepository.java
@@ -1,7 +1,0 @@
-package chatroom.repository;
-
-import chatroom.domain.Chatroom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChatroomRepository extends JpaRepository<Chatroom, Long> {
-}
