--- src/main/java/chatroom/repository/MessageRepository.java
@@ -1,7 +1,0 @@
-package chatroom.repository;
-
-import chatroom.domain.Message;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MessageRepository extends JpaRepository<Message, Long>  {
-}
