--- src/main/java/comment/repository/CommentReplyRepository.java
@@ -1,7 +1,0 @@
-package comment.repository;
-
-import comment.domain.CommentReply;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentReplyRepository extends JpaRepository<CommentReply, Long> {
-}
