--- src/main/java/comment/repository/CommentRepository.java
@@ -1,7 +1,0 @@
-package comment.repository;
-
-import comment.domain.Comment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CommentRepository extends JpaRepository<Comment, Long> {
-}
