--- src/main/java/post/domain/Category.java
@@ -1,24 +1,0 @@
-package post.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter @Setter
-public class Category {
-
-    @Id @GeneratedValue
-    @Column(name = "category_id")
-    private Long id;
-
-    private String name;
-
-    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL)
-    private List<PostCategory> postCategoryList = new ArrayList<PostCategory>();
-
-
-}
