--- src/main/java/post/repository/CategoryRepository.java
@@ -1,8 +1,0 @@
-package post.repository;
-
-import post.domain.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-
-}
