--- src/main/java/post/repository/PostCategoryRepository.java
@@ -1,7 +1,0 @@
-package post.repository;
-
-import post.domain.PostCategory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostCategoryRepository extends JpaRepository<PostCategory, Long> {
-}
