--- src/main/java/post/repository/PostRepository.java
@@ -1,8 +1,0 @@
-package post.repository;
-
-import post.domain.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-
-}
