+++ src/main/java/server/WOWMATE/Server/ServerApplication.java
@@ -1,0 +1,13 @@
+package server.WOWMATE.Server;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ServerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ServerApplication.class, args);
+	}
+
+}
