+++ src/main/java/server/WOWMATE/Server/test.java
@@ -1,0 +1,6 @@
+package server.WOWMATE.Server;
+
+public class test {
+
+
+}
