+++ src/main/java/server/chatroom/repository/ChatroomRepository.java
@@ -1,0 +1,7 @@
+package server.chatroom.repository;
+
+import server.chatroom.domain.Chatroom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatroomRepository extends JpaRepository<Chatroom, Long> {
+}
