+++ src/main/java/server/chatroom/repository/MessageRepository.java
@@ -1,0 +1,7 @@
+package server.chatroom.repository;
+
+import server.chatroom.domain.Message;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MessageRepository extends JpaRepository<Message, Long>  {
+}
