+++ src/main/java/server/comment/repository/CommentRepository.java
@@ -1,0 +1,7 @@
+package server.comment.repository;
+
+import server.comment.domain.Comment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
