+++ src/main/java/server/post/repository/CategoryRepository.java
@@ -1,0 +1,8 @@
+package server.post.repository;
+
+import server.post.domain.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+}
