+++ src/main/java/server/post/repository/PostCategoryRepository.java
@@ -1,0 +1,7 @@
+package server.post.repository;
+
+import server.post.domain.PostCategory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostCategoryRepository extends JpaRepository<PostCategory, Long> {
+}
