+++ src/main/java/server/user/domain/Gender.java
@@ -1,0 +1,5 @@
+package server.user.domain;
+
+public enum Gender {
+    M, F
+}
