+++ src/main/java/server/user/repository/SchoolRepository.java
@@ -1,0 +1,7 @@
+package server.user.repository;
+
+import server.user.domain.School;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SchoolRepository extends JpaRepository<School, Long> {
+}
