--- src/main/java/user/domain/Gender.java
@@ -1,5 +1,0 @@
-package user.domain;
-
-public enum Gender {
-    M, F
-}
