--- src/main/java/user/domain/School.java
@@ -1,24 +1,0 @@
-package user.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@Setter
-@ToString
-public class School {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "school_id")
-    private Long id;
-
-    private String name;
-
-    private String email_domain;
-
-}
