--- src/main/java/user/repository/SchoolRepository.java
@@ -1,7 +1,0 @@
-package user.repository;
-
-import user.domain.School;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SchoolRepository extends JpaRepository<School, Long> {
-}
